4.21

+++ controllers/blogs.js
@@ -23,6 +23,7 @@ blogsRouter.get('/', async (request, response) => {
     const body = request.body
 
     const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    console.log(decodedToken)
     if (!decodedToken.id) {
       return response.status(401).json({ error: 'token invalid' })
     }
@@ -46,9 +47,32 @@ blogsRouter.get('/', async (request, response) => {
     response.status(201).json(saveBlog)
   })
 
+  // blogsRouter.delete('/:id', async (request, response) => {
+  //   await Blog.findByIdAndDelete(request.params.id)
+  //   response.status(204).end()
+  // })
   blogsRouter.delete('/:id', async (request, response) => {
-    await Blog.findByIdAndDelete(request.params.id)
-    response.status(204).end()
+    const body = request.body
+
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!decodedToken.id) {
+      return response.status(401).json({ error: 'token invalid' })
+    }
+    const user = await User.findById(decodedToken.id)
+    if (!user) {
+      return response.status(401).json({ error: 'user invalid' })
+    }
+    const blog = await Blog.findById(request.params.id)
+    if (!blog) {
+      return response.status(401).json({ error: 'blog invalid' })
+    }
+    if ( blog.user.toString() === user.id.toString() ) {
+      await Blog.findByIdAndDelete(request.params.id)
+      response.status(204).end()
+    }
+    else{
+      response.status(401).json({ error: 'invalid operation' })
+    }
   })
 
   blogsRouter.put('/:id', async (request, response) => {
